Merged PR 5744: e4 deployment
This PR adds e4 into the deployment script.  Also does a minor cleanup on the CICD, forcing to be e1 only.

Below is the latest pipeline run, that worked as expected on e4 (with no exit gate):
https://dev.azure.com/bcz-prod/digital-verification-platform/_build/results?buildId=72304&view=results

When it got to e6, I cancelled on purpose.

Entry gates are setup on the environment itself, not through the pipeline code.  That is why you are not seeing any entry gate changes in the PR.

Thank you

Related work items: #21742